To support Windows the output lib was moved
As far as I understand the MSVC linker does not by default allow two libraries with the same name to be outputted to the same directory. This is what happens with Math.lib and the corresponding Math python library. To avoid this the output directory for targets created by tolc is now build/tolc